requirements.txt
flask
fpdf
requests
python-dotenv
